+++ src/layers/headers/CreditsLayer.h
@@ -4,6 +4,20 @@
 class CreditsLayer : public Popup<std::string const&> {
 protected:
     bool setup(std::string const& value) override;
+    void setupCreditsTab();
+    void setupLevelsTab();
+    void changeTab();
+    void onChangeTab(CCObject* sender);
+
+    CCMenuItemSpriteExtra* m_forwardTabBtn = nullptr;
+    CCMenuItemSpriteExtra* m_backTabBtn = nullptr;
+
+    CCMenu* m_forwardTabMenu = nullptr;
+    CCMenu* m_backTabMenu = nullptr;
+
+    CCLayer* m_creditsLayer = nullptr;
+    CCLayer* m_levelsLayer = nullptr;
+    int m_tab = 0;
 public:
     static CreditsLayer* create(std::string const& title);
 };