+++ changelog.md
@@ -1,4 +1,19 @@
 # Geometry Dash: Lunar Changelog
+## v2.0.0
+- New main levels "Power Glove", "Payback", "Reverie", "Ground Zero" and "Ultrasonic"
+- 4 platformer levels
+- New achievements
+- New icons
+- New vault
+- The "Throne Room"
+- New credits menu
+## v1.0.5
+- Added Lunar and Genamorphosis logo to loading screen
+- Changed how icons are replaced to make it much better
+- All achievements that are in the achievement menu are unlockable (no unlockable ones anymore)
+- Changed how the different save loading works
+- Updated all levels with changes made by Capeling
+- Fixed the level percentages being weird
 ## v1.0.4
 - Fixed Earthsplitter playing the incorrect song (no remix and remix were swapped)
 - Updated Earthsplitter